Jorm

Documentation for Jorm.

Jorm.RawSQL

Jorm.connect

Jorm.connect

Jorm.create_table

Jorm.delete!

Jorm.delete_all!

Jorm.disconnect

Jorm.disconnect

Jorm.execute_query

Jorm.execute_query

Jorm.filter_by

Jorm.filter_by_sql

Jorm.groupby_sql

Jorm.insert!

Jorm.list_tables

Jorm.ls_tables

Jorm.read_one

Jorm.tablename

Jorm.update!

Jorm.@raw_sql

Jorm.RawSQL — Type

struct RawSQL  
    value: String

A struct to hold raw SQL Queries

Jorm.connect — Method

connect(connection_string::String)
    -> LibPQ.Connection

Establish a connection to an SQLite database using the provided connection string.

Jorm.connect — Method

connect(connection_string::String)
    -> SQLite.DB

Establish a connection to an SQLite database using the provided connection string.

Jorm.create_table — Method

create_table(db::SQLite.DB, model, tableName)
Creates a table in the database based on the model.

Jorm.delete! — Method

delete!(db::SQLite.DB, model, id)
Delete data for a given model

Jorm.delete_all! — Method

delete_all(db::SQLite.DB, model) 
Deletes all data in the table for the given model

Jorm.disconnect — Method

disconnect(db::LibPQ.Connection)
    -> Nothing

Close the connection to the PostgreSQL database.

Jorm.disconnect — Method

disconnect(db::SQLite.DB)
    -> Nothing

Close the connection to the SQLite database.

Jorm.execute_query — Function

execute_query(db::LibPQ.Connection, query::RawSQL, params::Vector{Any}=Any[]) 
Returns the executed sql

Jorm.execute_query — Function

execute_query(db::SQLite.DB, query::RawSQL, params::Vector{Any}=Any[]) 
Returns the executed sql

Jorm.filter_by — Method

filter_by(db::SQLite.DB, model, table_name; kwargs...) 
Returns the result of the given condition from the Database. This uses `SELECT`

Jorm.filter_by_sql — Method

filter_by_sql(table_name; kwargs...)
Filter data by condition from the database

Jorm.groupby_sql — Method

groupby_sql(table_name; group_by_columns, select_columns = "*", having_conditions = nothing, order_by_columns = nothing) 
Returns the result of the given condition from the Database. This uses `SELECT`

Jorm.insert! — Method

insert!(db::SQLite.DB, model, data) 
Insert or Add data to the DB for a given model

Jorm.list_tables — Method

list_tables(connection_string::SQLiteConnectionString) 
Returns the list of tables in DB

Jorm.ls_tables — Method

ls_tables(connection_string::SQLiteConnectionString) 
Returns the list of tables in DB

Jorm.read_one — Method

read_one(db::SQLite.DB, model, id) 
Returns a given model object when given the ID

Jorm.tablename — Method

tablename(model) 
Generate the table name for a given model. The table name is derived from the model's type name, converting it to lowercase and joining words with underscores.

struct MyModel end println(tablename(MyModel)) # Output: "my_model"

Jorm.update! — Method

update!(db::SQLite.DB, model, id, data)
Update data to the DB for a given model

Jorm.@raw_sql — Macro

macro raw_sql(v::String)
    -> RawSQL

Create a RawSQL instance from a given SQL query string.